Set the default background and text color in order to make the front page legible on some configuration
This is needed when the visitor has changed his default browser `background-color` and `color` rules.
Those are usually `white` and `#333` by default, but some users change them to setup a 'dark theme',
which means than on the front page, if the font color `color` rule is set in the CSS, but the
`background-color` is not set, then you get a black-on-black result, making the text illegible.

Signed-off-by: Alexandre Bonneau <alexandre.bonneau@linuxfr.eu>